Fix primary key type in schema

+++ backend/prisma/migrations/20220412231554_init/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE `User` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT,
+    `id` INTEGER NOT NULL AUTO_INCREMENT,
     `email` VARCHAR(191) NOT NULL,
     `password` VARCHAR(191) NOT NULL,
     `financialCapital` DECIMAL(65, 30) NOT NULL,
@@ -10,8 +10,8 @@ CREATE TABLE `User` (
 
 -- CreateTable
 CREATE TABLE `Account` (
-    `account_id` BIGINT NOT NULL AUTO_INCREMENT,
-    `userId` BIGINT NULL,
+    `account_id` INTEGER NOT NULL AUTO_INCREMENT,
+    `userId` INTEGER NULL,
 
     PRIMARY KEY (`account_id`)
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;
